カラムの型を修正しました

+++ db/migrate/20210610040948_create_body_profiles.rb
@@ -2,8 +2,8 @@ class CreateBodyProfiles < ActiveRecord::Migration[6.1]
   def change
     create_table :body_profiles do |t|
       t.string :name, null: false
-      t.integer :height, null: false
-      t.integer :weight, null: false
+      t.float :height, null: false
+      t.float :weight, null: false
 
       t.timestamps
     end
