+++ db/schema.rb
@@ -14,8 +14,8 @@ ActiveRecord::Schema.define(version: 2021_06_10_040948) do
 
   create_table "body_profiles", force: :cascade do |t|
     t.string "name", null: false
-    t.integer "height", null: false
-    t.integer "weight", null: false
+    t.float "height", null: false
+    t.float "weight", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
